requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
langchain-ollama>=0.1.4
langchain-chroma>=0.1.1
chromadb>=0.5.5
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
pypdf>=4.3.1
huggingface-hub>=0.24.0
pytest>=8.2.0